RUE-1784: separate comptime dispatch from semantic hooks
